feat: rewrite parser for arrays, % and == ops
I tried to port the parser from python-osc more directly but had trouble
since they used lists as a stack for parsing the arrays.
Since the seqs in nims are less flexible I tried to write it in
a traidtional recursive way.
This works and passes all tests but the the way I pass i, j around
and the code in general looks a bit ugly I feel.

Also added % and == to compare the osc values, inspired by std/json.

Next step is gonna be the parsing of the osc type tag and then
probably bundles and serializing messages.