feat: implement and test message serialization
Next I need to test some malicious input and make sure it doesn't crash
and I want to do some randmoized comparison with python osc.

I am sure it is not optimal but all the test are passing.
Got the feeling that strings are not meant to be used as byte buffers
like this but I am not sure.