refactor: support for nimscript/comptime execution
This is a rather big commit.
Bascially I read some other code like flatly and decided to simplify
and abstract the reading/writing of bytes and integers in a way
that also supports nimscript/comptime execution.

The stream.nim contains the main code with functions that get used
by nosc.nim for the actual reading/writing of the osc data.

Additionally I added a bunch of utilities, benchmarks, improved
the test code a bit and improved the README(didn't need a template).